requirements.txt
gradio>=4.44.0
mlflow>=2.17.0
torch>=2.1.0
transformers>=4.45.0
peft>=0.13.0
groq>=0.11.0
python-dotenv>=1.0.0
openai>=1.50.0
requests>=2.31.0
boto3>=1.34.0
beautifulsoup4>=4.12.0
tiktoken>=0.7.0
youtube-transcript-api>=0.6.0
